fix(pr-review): address remaining PR feedback items
Split monolithic files and consolidate changelogs per review comments:

1. **war-room-specification.md** (2306 → 1092 lines)
   - Split into 7 modular spec files with cross-references
   - overview, architecture, experts, protocol, data, integration, implementation

2. **test_war_room_orchestrator.py** (2044 lines → 6 files)
   - merkle_dag, session, experts, phases, invocation, full_flow
   - All 86 tests passing

3. **Changelog consolidation**
   - Simplified attune/CHANGELOG.md to redirect to main
   - Simplified hookify/CHANGELOG.md to redirect to main
   - Added Hookify v1.1.0 and Attune v1.2.0 content to main CHANGELOG

4. **fix-pr workflow enhancements**
   - Added PENDING review state detection
   - Added troubleshooting for pending review issues

Addresses:
- #142 comment: split large spec file
- #142 comment: consolidate changelogs
- #142 comment: modularize test file
