+++ docs/specs/war-room-spec-implementation.md
@@ -1,0 +1,106 @@
+# War Room Specification: Implementation
+
+**Part of**: [War Room Specification](war-room-spec-overview.md)
+
+---
+
+## Implementation Phases
+
+### Phase 1: Foundation (Week 1)
+
+- [ ] Create `plugins/attune/skills/war-room/` structure
+- [ ] Implement basic `war_room_orchestrator.py`
+- [ ] Define expert prompt templates
+- [ ] Create Merkle-DAG data structures
+- [ ] Basic Strategeion directory structure
+
+**Deliverable**: Minimal working War Room with 3 experts
+
+### Phase 2: Deliberation Protocol (Week 2)
+
+- [ ] Implement all 7 phases
+- [ ] Add parallel dispatch via asyncio
+- [ ] Implement anonymization layer
+- [ ] Add voting aggregation
+- [ ] Premortem integration
+
+**Deliverable**: Full deliberation flow
+
+### Phase 3: Memory Palace Integration (Week 3)
+
+- [ ] Strategeion archive structure
+- [ ] Campaign history commands
+- [ ] Doctrine extraction skill
+- [ ] Session resume capability
+
+**Deliverable**: Persistent War Room with searchable history
+
+### Phase 4: Escalation & Refinement (Week 4)
+
+- [ ] Supreme Commander escalation logic
+- [ ] Delphi convergence mode
+- [ ] Meritocracy override
+- [ ] Hook for auto-trigger (gated)
+- [ ] Documentation and examples
+
+**Deliverable**: Production-ready War Room
+
+---
+
+## Success Criteria
+
+### Functional
+
+- [ ] Can invoke War Room with problem statement
+- [ ] Experts respond in parallel when possible
+- [ ] Anonymization masks attribution until unsealed
+- [ ] Red Team challenges all COAs
+- [ ] Premortem identifies failure modes
+- [ ] Supreme Commander synthesizes final decision
+- [ ] Session persists to Strategeion
+
+### Quality
+
+- [ ] Deliberation improves decision quality vs single-model
+- [ ] Diverse perspectives surface non-obvious risks
+- [ ] Red Team catches at least 3 assumptions per COA
+- [ ] Premortem identifies realistic failure modes
+
+### Performance
+
+- [ ] Lightweight mode completes in < 5 minutes
+- [ ] Full council completes in < 15 minutes
+- [ ] Graceful degradation when expert fails
+
+### Integration
+
+- [ ] Works with existing conjure delegation
+- [ ] Integrates with memory-palace architecture
+- [ ] Can be invoked from brainstorm workflow
+- [ ] Session state is resumable
+
+---
+
+## References
+
+### Strategic Frameworks
+- [Sun Tzu - Art of War](https://classics.mit.edu/Tzu/artwar.html)
+- [Clausewitz - On War](https://clausewitzstudies.org/readings/OnWar1873/BK1ch07.html)
+- [Robert Greene - 33 Strategies](https://grahammann.net/book-notes/the-33-strategies-of-war-robert-greene)
+- [MDMP - U.S. Army](https://www.army.mil/article/271773/military_decision_making_process_organizing_and_conducting_planning)
+
+### Decision Methods
+- [Gary Klein - Premortem](https://www.gary-klein.com/premortem)
+- [Delphi Method](https://en.wikipedia.org/wiki/Delphi_method)
+- [Six Thinking Hats](https://www.debonogroup.com/services/core-programs/six-thinking-hats/)
+
+### Multi-Agent Systems
+- [Karpathy LLM Council](https://github.com/karpathy/llm-council)
+- [MIT Multi-AI Collaboration](https://news.mit.edu/2023/multi-ai-collaboration-helps-reasoning-factual-accuracy-language-models-0918)
+- [Multi-Agent Debate](https://arxiv.org/abs/2305.14325)
+- [Society of Mind](https://en.wikipedia.org/wiki/Society_of_Mind)
+
+---
+
+**Status**: Draft - Awaiting Review
+**Next**: Implementation Phase 1 upon approval
